refactor: follows underscore naming
