+++ src/sns_wasm.ts
@@ -6,7 +6,7 @@ import {
   SnsWasmService,
 } from "../candid/sns_wasm.idl";
 import { MAINNET_SNS_WASM_CANISTER_ID } from "./constants/canister_ids";
-import { SnsWasmCanisterOptions } from "./types/sns-wasm";
+import { SnsWasmCanisterOptions } from "./types/sns_wasm";
 import { defaultAgent } from "./utils/agent.utils";
 
 export class SnsWasmCanister {
